pyproject.toml
[tool.poetry]
name = "octobotv4"
version = "5"
description = "Asyncio version of octobot"
authors = ["Yan <me@octonezd.me>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "20.0a2"
aioredis = "^2.0.1"
aiohttp = "^3.8.1"
pyngrok = "^5.1.0"
aiohttp-client-cache = "^0.7.1"
fakeredis = "^1.8.1"
Babel = "^2.10.3"
emoji-country-flag = "^1.3.0"
toml = "^0.10.2"
uvicorn = "^0.18.2"
fastapi = "^0.78.0"
psutil = "^5.9.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
pygenstub = "^2.0.1"
uvicorn = {extras = ["standard"], version = "^0.18.2"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
